pyproject.toml
[tool.poetry]
name = "code-bench"
version = "0.1.0"
description = ""
authors = ["Vladislav Kirienko ve.kirienko7@gmail.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pygithub = "^2.5.0"
mypy = "^1.13.0"
black = {extras = ["jupyter"], version = "^24.10.0"}
pylint = "^3.3.1"
flake8 = "^7.1.1"
tqdm = "^4.67.0"
isort = "^5.13.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
